+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-### Unreleased
+### v1.0.0
 
 ##### Added
 
 * Implement "a version" of Recipients Interceptor https://github.com/croaky/recipient_interceptor compatible with SendGrid.
 * Raise exceptions when api fails.
 
-##### Changed
+##### Removed
 
 * Remove ability to set templates by name